+++ src/components/CreateButton.tsx
@@ -244,13 +244,8 @@ export const CreateButton = () => {
                     setBolt12Offer(undefined);
                     setInvoiceValid(true);
                 } catch (e) {
-                    const err: unknown =
-                        typeof e.json === "function"
-                            ? (await e.json()).error
-                            : e;
-
-                    notify("error", formatError(err));
-                    log.warn("Fetching invoice from bol12 failed", err);
+                    notify("error", formatError(e));
+                    log.warn("Fetching invoice from bol12 failed", e);
                     return;
                 }
             }
@@ -362,24 +357,19 @@ export const CreateButton = () => {
                 isRecklessMode() ||
                 swapType() === SwapType.Reverse ||
                 assetSend() === RBTC ||
-                isMobileEvmBrowser()
+                // Only disable refund files on mobile EVM browsers when one side is RSK
+                (assetReceive() === RBTC && isMobileEvmBrowser())
             ) {
                 navigate("/swap/" + data.id);
             } else {
                 navigate("/swap/refund/" + data.id);
             }
         } catch (err) {
-            let msg = err;
-
-            if (typeof err.json === "function") {
-                msg = (await err.json()).error;
-            }
-
-            if (msg === "invalid pair hash") {
+            if (err === "invalid pair hash") {
                 setPairs(await getPairs());
                 notify("error", t("feecheck"));
             } else {
-                notify("error", msg);
+                notify("error", err);
             }
         }
     };