+++ src/i18n/i18n.ts
@@ -40,7 +40,7 @@ const dict = {
         minimum_amount: "Minimum amount is {{ amount }} {{ denomination }}",
         maximum_amount: "Maximum amount is {{ amount }} {{ denomination }}",
         assets: "Multi-Layer",
-        footer: "Made with ❤️ by Team Boltz",
+        footer: "Made with 🧡 by Team Boltz",
         create_channel: "Create Lightning Channel",
         create_channel_subline: "Channel Inbound or Outbound Capacity",
         create_swap: "Create Atomic Swap",
@@ -96,18 +96,29 @@ const dict = {
         backup_refund_list_clear_history: "clear your browsing history",
         backup_refund_skip:
             "If any of the above applies to you, we strongly recommend downloading this refund file!",
-        refund_a_swap: "Refund a swap",
-        refund_swap: "Refund Swap {{ id }}",
+        refund_swap: "Refund Swap",
         refund_a_swap_subline:
-            "Upload your refund file and reclaim your locked funds",
+            "Upload your Swap’s refund file, to refund a swap that is not available in this browser’s swap history.",
         refund_past_swaps: "Past swaps",
         refund_past_swaps_subline:
             "Swaps that got saved into your browsers storage",
+        no_refundable_swaps:
+            "No refundable swaps found in your browser history",
+        cant_find_swap: "Can't find your swap?",
+        refund_external_explainer:
+            "Try refunding an external swap via refund file and other emergency methods",
+        refund_external_explainer_rsk:
+            "Connect your Rootstock Wallet to scan for refundable swaps that are not saved in this browser’s swap history.",
+        refund_external_scanning_rsk:
+            "Scanning for refundable swaps in your Rootstock Wallet...",
+        connected_wallet_no_swaps:
+            "The connected Rootstock Wallet does not contain any refundable swaps.",
+        refund_external_swap: "Refund External Swap",
         history_no_swaps: "Looks like you didn't do any swaps yet.",
         refund_address_header:
-            "Enter address of your {{ asset }} wallet to refund",
+            "Enter a {{ asset }} address to receive your refund on:",
         refund_address_header_no_asset:
-            "Enter address of your wallet to refund",
+            "Enter an address to receive your refund on:",
         refund_backup: "Backup",
         refund_import: "Import Backup",
         refund_clear: "Delete storage",
@@ -157,6 +168,8 @@ const dict = {
         peers: "Number of Peers",
         num_channels: "Number of Channels",
         beta_caution: "BETA - USE WITH CAUTION!",
+        pro: "pro",
+        pro_banner: "BETA: AVAILABILITY NOT GUARANTEED",
         warning_return: "Important: Return to this page after paying invoice",
         warning_expiry:
             "Make sure your transaction confirms within ~24 hours after creation of this swap!",
@@ -172,8 +185,6 @@ const dict = {
         version: "Version",
         commithash: "Commit Hash",
         open_in_wallet: "Open in Wallet",
-        invoice_paid: "Invoice paid",
-        creating_coop_claim: "Creating cooperative claim transaction.",
         branding: "Branding",
         testnet: "Testnet",
         broadcasting_claim: "Broadcasting claim transaction...",
@@ -226,6 +237,9 @@ const dict = {
         sent: "Sent",
         will_receive: "Will receive",
         refund_available_in: "Refund will be available in {{ blocks }} blocks",
+        no_wallet_connected: "No wallet connected",
+        no_lockup_transaction: "No lockup transaction found",
+        routing_fee_limit: "Routing fee limit",
     },
     de: {
         language: "Deutsch",
@@ -269,7 +283,7 @@ const dict = {
         minimum_amount: "Mindestbetrag ist {{ amount }} {{ denomination }}",
         maximum_amount: "Höchstbetrag ist {{ amount }} {{ denomination }}",
         assets: "Multi-layer",
-        footer: "Gemacht mit ❤️ von Team Boltz",
+        footer: "Mit 🧡 gemacht von Team Boltz",
         create_channel: "Erstelle Lightning-Kanal",
         create_channel_subline: "Eingehende oder ausgehende Kapazität",
         create_swap: "Erstelle Atomic Swap",
@@ -326,13 +340,24 @@ const dict = {
         backup_refund_list_clear_history: "die Browser Daten gelöscht werden",
         backup_refund_skip:
             "Falls einer dieser Punkte zutrifft, empfehlen wir ausdrücklich die Rückerstattungsdatei herunterzuladen!",
-        refund_a_swap: "Einen Swap erstatten",
-        refund_swap: "Erstatte Swap {{ id }}",
+        refund_swap: "Swap Erstatten",
         refund_a_swap_subline:
-            "Lade deine Rückerstattungsdatei hoch und hole dir deine Bitcoin aus einem fehlgeschlagenen Swap zurück",
+            "Lade deine Rückerstattungsdatei hoch und hole dir deine Bitcoin aus einem fehlgeschlagenen Swap zurück.",
         refund_past_swaps: "Historische Swaps",
         refund_past_swaps_subline:
             "Swaps, die in deinem Browser gespeichert wurden",
+        no_refundable_swaps:
+            "Keine erstattungsfähigen Swaps in Browserverlauf gefunden",
+        cant_find_swap: "Swap nicht gefunden?",
+        refund_external_explainer:
+            "Versuchen Sie, einen externen Swap über eine Rückerstattungsdatei und andere Notfallmethoden zu erstatten",
+        refund_external_explainer_rsk:
+            "Verbinden Sie Ihr Rootstock Wallet, um nach erstattungsfähigen Swaps zu suchen, die nicht im Swap-Verlauf dieses Browsers gespeichert sind..",
+        refund_external_scanning_rsk:
+            "Scanne nach erstattungsfähigen Swaps in Rootstock-Wallet...",
+        connected_wallet_no_swaps:
+            "Das angeschlossene Rootstock Wallet enthält keine erstattungsfähigen Swaps.",
+        refund_external_swap: "Externen Swap rückerstatten",
         history_no_swaps:
             "Es sieht so aus, als hättest du noch nicht geswappt.",
         refund_address_header: "Adresse deines {{ asset }} Wallets eingeben",
@@ -390,6 +415,8 @@ const dict = {
         peers: "Anzahl der Peers",
         num_channels: "Anzahl der Kanäle",
         beta_caution: "BETA - OBACHT!",
+        pro: "pro",
+        pro_banner: "BETA: VERFÜGBARKEIT NICHT GARANTIERT",
         warning_return:
             "Wichtig: Kehre nach dem Bezahlen der Rechnung zu dieser Seite zurück!",
         warning_expiry:
@@ -406,8 +433,6 @@ const dict = {
         version: "Version",
         commithash: "Commit Hash",
         open_in_wallet: "Im Wallet öffnen",
-        invoice_paid: "Rechnung bezahlt",
-        creating_coop_claim: "Erstelle kooperative Claim-Transaktion.",
         branding: "Branding",
         testnet: "Testnet",
         broadcasting_claim: "Sende claim transaction...",
@@ -464,6 +489,9 @@ const dict = {
         sent: "Gesendet",
         will_receive: "Sie erhalten",
         refund_available_in: "Rückerstattung möglich in {{ blocks }} Blöcken",
+        no_wallet_connected: "Kein Wallet verbunden",
+        no_lockup_transaction: "Keine Lockup-Transaktion gefunden",
+        routing_fee_limit: "Routing Gebühr Limit",
     },
     es: {
         language: "Español",
@@ -506,7 +534,7 @@ const dict = {
         minimum_amount: "La cantidad mínima es {{ amount }} {{ denomination }}",
         maximum_amount: "La cantidad máxima es {{ amount }} {{ denomination }}",
         assets: "Multicapa",
-        footer: "Hecho con ❤️ por el equipo de Boltz",
+        footer: "Hecho con 🧡 por el equipo de Boltz",
         create_channel: "Crear canal Lightning",
         create_channel_subline: "Capacidad del canal entrante o saliente",
         create_swap: "Crear intercambio atómico",
@@ -561,13 +589,24 @@ const dict = {
         backup_refund_list_clear_history: "borra tu historial de navegación",
         backup_refund_skip:
             "Si alguno de los puntos anteriores le afecta, le recomendamos que descargue el archivo de reembolso.",
-        refund_a_swap: "Reembolsar un intercambio",
-        refund_swap: "Reembolsar intercambio {{ id }}",
+        refund_swap: "Reembolsar Intercambio",
         refund_a_swap_subline:
-            "Cargue su archivo de reembolso y recupere sus fondos bloqueados",
+            "Cargue su archivo de reembolso y recupere sus fondos bloqueados.",
         refund_past_swaps: "Intercambios anteriores",
         refund_past_swaps_subline:
             "Intercambios que se guardaron en el almacenamiento del navegador",
+        no_refundable_swaps:
+            "No se han encontrado swaps reembolsables en el historial de tu navegador",
+        cant_find_swap: "¿No encuentra su intercambio?",
+        refund_external_explainer:
+            "Intenta reembolsar un swap externo mediante el archivo de reembolso y otros métodos de emergencia.",
+        refund_external_explainer_rsk:
+            "Conecta tu monedero Rootstock para buscar swaps reembolsables que no estén guardados en el historial de swaps de este navegador.",
+        refund_external_scanning_rsk:
+            "Escaneando en busca de swaps reembolsables en su monedero Rootstock...",
+        connected_wallet_no_swaps:
+            "El monedero Rootstock conectada no contiene ningún swap reembolsable.",
+        refund_external_swap: "Reembolsar Swap Externo",
         history_no_swaps: "Parece que aún no has realizado ningún intercambio.",
         refund_address_header:
             "Introduzca la dirección de tu monedero {{ asset }} para reembolsar",
@@ -626,6 +665,8 @@ const dict = {
         peers: "Número de peers",
         num_channels: "Número de canales",
         beta_caution: "BETA - ¡ÚSALO CON PRECAUCIÓN!",
+        pro: "pro",
+        pro_banner: "BETA: DISPONIBILIDAD NO GARANTIZADA",
         warning_return:
             "Importante: Regresa a esta página después de pagar la factura!",
         warning_expiry:
@@ -642,9 +683,6 @@ const dict = {
         version: "Versión",
         commithash: "Commit Hash",
         open_in_wallet: "Abrir en monedero",
-        invoice_paid: "Factura pagada",
-        creating_coop_claim:
-            "Creación de una transacción de reclamación cooperativa.",
         branding: "Branding",
         testnet: "Testnet",
         broadcasting_claim: "Enviando transacción de reclamación...",
@@ -699,6 +737,9 @@ const dict = {
         sent: "Enviado",
         will_receive: "Recibirá",
         refund_available_in: "Reembolso disponible en {{ blocks }} bloques",
+        no_wallet_connected: "No hay monedero conectado",
+        no_lockup_transaction: "No se encontró ninguna transacción de lockup",
+        routing_fee_limit: "Límite de la tarifa de enrutamiento",
     },
     zh: {
         language: "中文",
@@ -741,7 +782,7 @@ const dict = {
         minimum_amount: "最小金额为{{ amount }}{{ denomination }}",
         maximum_amount: "最大金额为{{ amount }}{{ denomination }}",
         assets: "多层",
-        footer: "Boltz团队倾情 ❤️ 制作",
+        footer: "Boltz团队倾情🧡制作",
         create_channel: "创建闪电网络通道",
         create_channel_subline: "通道入站或出站容量",
         create_swap: "创建原子交换",
@@ -792,11 +833,21 @@ const dict = {
         backup_refund_list_tor: "正在使用Tor浏览器",
         backup_refund_list_clear_history: "清除您的浏览历史记录",
         backup_refund_skip: "如果您符合以上任何条件，强烈建议下载此退款文件！",
-        refund_a_swap: "退还交换",
-        refund_swap: "退交换{{ id }}",
+        refund_swap: "退还交换",
         refund_a_swap_subline: "上传您的退款文件，取回被锁定的资金",
         refund_past_swaps: "过去的交换",
         refund_past_swaps_subline: "保存在浏览器存储中的交换",
+        no_refundable_swaps: "在您的浏览器历史中未发现可退款的交换。",
+        cant_find_swap: "找不到您的交换？",
+        refund_external_explainer:
+            "尝试通过退款文件和其他紧急方法退款外部交换。",
+        refund_external_explainer_rsk:
+            "连接您的 Rootstock 钱包，扫描未保存在此浏览器交换历史记录中的可退款交换。",
+        refund_external_scanning_rsk:
+            "正在扫描您的 Rootstock 钱包中的可退款掉期。。。",
+        connected_wallet_no_swaps:
+            "已连接的 Rootstock 钱包不包含任何可退款掉期。",
+        refund_external_swap: "退款外部交换",
         history_no_swaps: "看起来您还没有进行任何交换。",
         refund_address_header: "输入要退款的 {{ asset }} 钱包地址",
         refund_address_header_no_asset: "输入要退款的钱包地址",
@@ -845,6 +896,8 @@ const dict = {
         peers: "对端数",
         num_channels: "通道数",
         beta_caution: "BETA - 谨慎使用！",
+        pro: "pro",
+        pro_banner: "测试版：不保证可用性",
         warning_return: "重要：支付发票后返回此页面",
         warning_expiry: "确保您的交易在创建此交换后的约24小时内确认！",
         not_found: "404 - 页面未找到",
@@ -859,8 +912,6 @@ const dict = {
         version: "版本",
         commithash: "提交哈希",
         open_in_wallet: "在钱包中打开",
-        invoice_paid: "发票已支付",
-        creating_coop_claim: "创建合作索赔交易。",
         branding: "品牌",
         testnet: "测试网",
         broadcasting_claim: "正在发送索赔交易...",
@@ -907,6 +958,9 @@ const dict = {
         sent: "已发送",
         will_receive: "将收到",
         refund_available_in: "退款将分 {{ blocks }} 区块提供",
+        no_wallet_connected: "未连接钱包",
+        no_lockup_transaction: "未找到锁仓交易",
+        routing_fee_limit: "最大路由费用",
     },
     ja: {
         language: "日本語",
@@ -949,7 +1003,7 @@ const dict = {
         minimum_amount: "最小金額は{{ amount }} {{ denomination }}です",
         maximum_amount: "最大金額は{{ amount }} {{ denomination }}です",
         assets: "アセット",
-        footer: "Team Boltzによる❤️で作成されました",
+        footer: "Team Boltzによる🧡で作成されました",
         create_channel: "Lightningチャネルの作成",
         create_channel_subline:
             "チャネルのインバウンドまたはアウトバウンド容量",
@@ -1009,12 +1063,23 @@ const dict = {
         backup_refund_list_clear_history: "ブラウザ履歴を消した場合",
         backup_refund_skip:
             "もし以上にあてはまる場合、返金ファイルをダウンロードすることを強くお勧めします！",
-        refund_a_swap: "スワップを返金する",
-        refund_swap: "スワップの返金 {{ id }}",
+        refund_swap: "スワップを返金する",
         refund_a_swap_subline:
             "返金ファイルをアップロードし、ロックされた資金を回収します",
         refund_past_swaps: "過去のスワップ",
         refund_past_swaps_subline: "ブラウザのストレージに保存されたスワップ",
+        no_refundable_swaps:
+            "ブラウザの履歴に返金可能なスワップが見つかりません。",
+        cant_find_swap: "スワップが見つからない？",
+        refund_external_explainer:
+            "払い戻しファイルやその他の緊急手段を使って外部スワップを払い戻してみる。",
+        refund_external_explainer_rsk:
+            "Rootstockウォレットを接続して、このブラウザのスワップ履歴に保存されていない払い戻し可能なスワップをスキャンします。",
+        refund_external_scanning_rsk:
+            "Rootstockウォレットにある払い戻し可能なスワップをスキャンしています...",
+        connected_wallet_no_swaps:
+            "接続されているRootstockウォレットには返金可能なスワップが含まれていません。",
+        refund_external_swap: "払い戻し外部スワップ",
         history_no_swaps: "まだスワップを行っていないようです。",
         refund_address_header:
             "返金用の {{ asset }} ウォレットのアドレスを入力",
@@ -1070,6 +1135,8 @@ const dict = {
         peers: "ピアの数",
         num_channels: "チャネルの数",
         beta_caution: "ベータ版の為、ご利用は慎重にお願いします",
+        pro: "pro",
+        pro_banner: "ベータ版：利用可能性は保証されない",
         warning_return:
             "重要：インボイスのお支払い後、このページへ戻ってください",
         warning_expiry:
@@ -1086,8 +1153,6 @@ const dict = {
         version: "バージョン",
         commithash: "コミットハッシュ",
         open_in_wallet: "ウォレットで開く",
-        invoice_paid: "インボイス支払い完了",
-        creating_coop_claim: "協力的クレームトランザクションの作成中",
         branding: "ブランディング",
         testnet: "テストネット",
         broadcasting_claim: "クレームトランザクションのブロードキャスト中",
@@ -1140,6 +1205,9 @@ const dict = {
         sent: "送信済み",
         will_receive: "受信予定",
         refund_available_in: "返金は {{ blocks }} つのブロックに分かれる",
+        no_wallet_connected: "財布はつながっていない！",
+        no_lockup_transaction: "ロックアップトランザクションが見つかりません",
+        routing_fee_limit: "ルーティング料金の上限",
     },
 };