+++ src/utils/boltzClient.ts
@@ -35,8 +35,9 @@ type SubmarinePairTypeTaproot = PairType & {
         maximalZeroConf: number;
     };
     fees: {
-        percentage: number;
         minerFees: number;
+        percentage: number;
+        maximalRoutingFee?: number;
     };
 };
 
@@ -416,9 +417,9 @@ export const getChainSwapClaimDetails = (id: string) =>
 
 export const postChainSwapDetails = (
     id: string,
-    preimage: string,
+    preimage: string | undefined,
     signature: { pubNonce: string; partialSignature: string },
-    toSign: { pubNonce: string; transaction: string; index: number },
+    toSign?: { pubNonce: string; transaction: string; index: number },
 ) => {
     checkCooperative();
     return fetcher<{